+++ AudioPlayer/AudioPlayer/Content/AudioQueue/AudioQueueView.swift
@@ -6,11 +6,22 @@
 import SwiftUI
 
 struct AudioQueueView: View {
+    @Environment(AppModel.self) var appModel
+
+    @State var model: AudioPlayerModel
+
+    @State var eqSheetIsShown: Bool = false
+    @State var addNewAudioIsShown: Bool = false
+
+    init(appModel: AppModel) {
+        self._model = State(wrappedValue: AudioPlayerModel(audioPlayerService: appModel.audioPlayerService))
+    }
+    
     var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+        return Text("")
     }
 }
 
 #Preview {
-    AudioQueueView()
+    AudioQueueView(appModel: AppModel())
 }
